requirements.txt
anthropic==0.49.0
colorama==0.4.6
dogstatsd_python==0.5.6
fastapi==0.115.12
fastapi_limiter==0.1.6
fasttext==0.9.3
httpx==0.28.1
numpy==2.2.5
prometheus_client==0.21.1
protobuf==6.30.2
psutil==7.0.0
psycopg2==2.9.10
pydantic==2.11.3
pydantic_settings==2.9.1
PyJWT==2.10.1
pytest==8.3.5
python-dotenv==1.1.0
python_bcrypt==0.3.2
python_dateutil==2.9.0.post0
PyYAML==6.0.2
redis==5.2.1
Requests==2.32.3
scikit_learn==1.6.1
sentry_sdk==2.26.1
spacy==3.8.5
starlette==0.46.2
torch==2.6.0
transformers==4.51.3
uvicorn==0.30.2
# Authentication and security dependencies
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
# Middleware and monitoring
starlette-admin==0.8.0
starlette-context==0.3.6
starlette-csrf>=1.5.0
uvicorn[standard]>=0.22.0